# openapi.yaml
openapi: 3.0.0
info:
  title: API Locação de Veiculos
  description: API para gerenciar carros, usuários e reservas.
  version: 1.0.0
  contact:
    name: Eduardo Ferreira
    email: eduardoferreirabmatos@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Servidor Local
components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: integer
          example: 1
        model:
          type: string
          example: "Tesla Model S"
        color:
          type: string
          example: "Red"
        year:
          type: integer
          example: 2022
        valuePerDay:
          type: number
          format: float
          example: 150.0
        numberOfPassengers:
          type: integer
          example: 5
        acessories:
          type: array
          items:
            type: string
          example: ["GPS", "Air Conditioner", "Sunroof"]
        reserves:
          type: array
          items:
            $ref: '#/components/schemas/Reserve'
        createdAt:
          type: string
          format: date-time
          example: "2023-10-08T15:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-08T16:30:00Z"
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        cpf:
          type: string
          example: "123.456.789-00"
        birth:
          type: string
          format: date
          example: "1990-01-01"
        qualified:
          type: boolean
          example: true
        cep:
          type: string
          example: "12345-678"
        neighbordhood:
          type: string
          example: "Downtown"
        street:
          type: string
          example: "123 Main St"
        complement:
          type: string
          example: "Apartment 301"
        city:
          type: string
          example: "São Paulo"
        uf:
          type: string
          example: "SP"
        email:
          type: string
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          example: "strongpassword123"
        reserves:
          type: array
          items:
            $ref: '#/components/schemas/Reserve'
        createdAt:
          type: string
          format: date-time
          example: "2023-10-08T15:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-08T16:30:00Z"
...
    Reserve:
      type: object
      properties:
        id:
          type: integer
          example: 1
        startDate:
          type: string
          format: date
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          example: "2024-01-10"
        finalValues:
          type: number
          format: float
          example: 1500.0
        user:
          $ref: '#/components/schemas/User'
        car:
          $ref: '#/components/schemas/Car'
        createdAt:
          type: string
          format: date-time
          example: "2023-10-08T15:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-08T16:30:00Z"
...
paths:
...
  /v1/reserve/{id}:
    get:
      summary: Retorna uma reserva específica
      tags:
        - Reserve
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID da reserva
      responses:
        '200':
          description: Detalhes da reserva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserve'
        '404':
          description: Reserva não encontrada

    delete:
      summary: Remove uma reserva
      tags:
        - Reserve
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID da reserva
      responses:
        '204':
          description: Reserva removida com sucesso
        '404':
          description: Reserva não encontrada

    put:
      summary: Atualiza uma reserva
      tags:
        - Reserve
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID da reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reserve'
      responses:
        '200':
          description: Reserva atualizada com sucesso
        '404':
          description: Reserva não encontrada